[FIX] Portado o modulo do cashflow report para o modulo do financeiro
