--- cashflow_report/__init__.py
@@ -1,2 +1,0 @@
-from . import reports
-from . import wizards
