--- cashflow_report/reports/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2017 KMEE
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import abstract_report_xlsx
-from . import financial_cashflow_report
-from . import financial_cashflow_report_xlsx
