--- cashflow_report/wizards/financial_cashflow_wizard_view.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <!-- WIZARD FINANCIAL CASHFLOW -->
-    <record id="trial_balance_wizard" model="ir.ui.view">
-        <field name="name">Report Cashflow</field>
-        <field name="model">trial.balance.report.wizard</field>
-        <field name="arch" type="xml">
-            <form>
-                <group name="main_info">
-                    <field name="company_id" options="{'no_create': True}" groups="base.group_multi_company"/>
-                </group>
-                <group string="Period Info">
-                    <field name="period"/>
-                </group>
-                <group>
-                    <group>
-                        <field name="date_from"/>
-                    </group>
-                    <group>
-                        <field name="date_to"/>
-                    </group>
-                    <group>
-                        <field name="hide_account_balance_at_0"/>
-                    </group>
-                </group>
-                <footer>
-                    <button name="button_export_xlsx" string="Export XLSX"
-                            type="object"/>
-                    <button string="Cancel" class="oe_link" special="cancel"/>
-                </footer>
-            </form>
-        </field>
-    </record>
-
-    <act_window id="action_trial_balance_wizard"
-                name="Report Cashflow"
-                res_model="trial.balance.report.wizard"
-                view_type="form"
-                view_mode="form"
-                view_id="trial_balance_wizard"
-                target="new" />
-
-    <record model="ir.ui.menu" id="financial_trial_balance">
-        <field name="name">Cashflow</field>
-        <field name="parent_id" ref="financial.financial_report_menu"/>
-        <field name="action" ref="action_trial_balance_wizard"/>
-        <field name="sequence" eval="19"/>
-    </record>
-
-</odoo>
