+++ financial/__init__.py
@@ -1,3 +1,3 @@
 from . import models
 from . import wizards
-# from . import report
+from . import report
