+++ financial/report/__init__.py
@@ -2,5 +2,8 @@
 # Copyright 2017 KMEE
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import financial_cashflow
-from . import financial_statement_report
+# from . import financial_cashflow
+# from . import financial_statement_report
+from. import abstract_report_xlsx
+from. import financial_cashflow_report
+from. import financial_cashflow_report_xlsx
