+++ financial/wizards/__init__.py
@@ -1,6 +1,6 @@
 # from . import financial_cancel
 # from . import financial_edit
 # from . import financial_create
-from . import financial_cashflow
 # from . import financial_pay_receive
 # from . import financial_renegociate
+from . import financial_cashflow
