+++ financial/wizards/financial_cashflow.xml
@@ -1,52 +1,51 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2017 KMEE
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-
 <odoo>
-<data>
 
-    <record model="ir.ui.view" id="financial_wizard_cashflow_form_view">
-        <field name="name">financial.wizard.cashflow.form (in financial)</field>
-        <field name="model">wizard.financial.cashflow</field>
+    <!-- WIZARD FINANCIAL CASHFLOW -->
+    <record id="trial_balance_wizard" model="ir.ui.view">
+        <field name="name">Report Cashflow</field>
+        <field name="model">trial.balance.report.wizard</field>
         <field name="arch" type="xml">
-            <form string="Financial Cashflow">
-                <!-- TODO -->
+            <form>
+                <group name="main_info">
+                    <field name="company_id" options="{'no_create': True}" groups="base.group_multi_company"/>
+                </group>
+                <group string="Period Info">
+                    <field name="period"/>
+                </group>
                 <group>
                     <group>
-                        <field name="period"/>
                         <field name="date_from"/>
+                    </group>
+                    <group>
                         <field name="date_to"/>
+                    </group>
+                    <group>
+                        <field name="hide_account_balance_at_0"/>
                     </group>
                 </group>
                 <footer>
-                    <button name='doit'
-                        string='OK'
-                        class='btn-primary oe_highlight'
-                        type='object'/>
-                    or
-                    <button string='Cancel'
-                        class='btn-default oe_link'
-                        special='cancel'/>
+                    <button name="button_export_xlsx" string="Export XLSX"
+                            type="object"/>
+                    <button string="Cancel" class="oe_link" special="cancel"/>
                 </footer>
             </form>
         </field>
     </record>
 
-    <record model="ir.actions.act_window" id="financial_cashflow_form_act_window">
-        <field name="name">Financial Cashflow Form</field>
-        <field name="res_model">wizard.financial.cashflow</field>
-        <field name="view_mode">form</field>
-        <field name="context">{}</field>
-        <field name="target">new</field>
+    <act_window id="action_trial_balance_wizard"
+                name="Report Cashflow"
+                res_model="trial.balance.report.wizard"
+                view_type="form"
+                view_mode="form"
+                view_id="trial_balance_wizard"
+                target="new" />
+
+    <record model="ir.ui.menu" id="financial_trial_balance">
+        <field name="name">Cashflow</field>
+        <field name="parent_id" ref="financial.financial_report_menu"/>
+        <field name="action" ref="action_trial_balance_wizard"/>
+        <field name="sequence" eval="19"/>
     </record>
 
-    <record model="ir.ui.menu" id="financial_cashflow_form_menu">
-        <field name="name">Cashflow Form</field>
-        <field name="parent_id" ref="financial_report_menu"/>
-        <field name="action" ref="financial_cashflow_form_act_window"/>
-        <field name="sequence" eval="18"/>
-    </record>
-
-
-</data>
 </odoo>
